Fix `load_context` shouldn't change current working directory (#489)
## tl;dr

This commit adds a temporary solution that makes logging and dataset work out of the box without changing the current working directory in `load_context`. Specifically, after loading configuration from `logging.yml` and `catalog.yml`, we recursively traverse through the config dictionary, attempt to expand any relative path we encounter for certain keys (e.g. `filepath`, `path`, `filename`) with absolute path, before passing the configuration on for further processing.

## 5-whys (ish) analysis of the original problem

### 1. Why is changing the current working directory to  project root  in `load_context` a problem?

The reason is 2-fold:

- Concretely, some workflows, e.g. IPython notebook, streamlit, etc., don't go through `run.py` or `kedro run` entrypoint. In other words, they are not run from the project root. Therefore, changing the current working directory back to the project root causes unexpected behaviour for these workflows.
- Conceptually, causing unexpected, undocumented side-effect on a user's filesystem is generally considered bad practice.

### 2. Why does Kedro need to change the current working directory inside `load_context` to be the project root

There are 2 reasons:

- We support specifying local dataset's file paths in `data/` using relative paths, e.g. `data/01_raw/companies.csv`
- The configuration for logging in a project, e.g. `logs/error.log` , uses relative paths.

Therefore, we have to change the current working directory to be the project root for these paths to resolve correctly.

### 3. Why do configuration for logging and local dataset use relative path instead of absolute path

Because a kedro project runs on different environments (different machines, different infrastructure, etc.), where the absolute path will be different.

## Problems

Based on the 5-why analysis above, the problem can be stated as: how can we keep a configuration API (for both logging and dataset) based on relative path but resolve to absolute path in `KedroContext` execution timeline without changing the current working directory?

If we frame the problem this way, the only solution is: there needs to be some extra processing after loading configuration from the YML files and before passing it on to `logging` and `DataCatalog`. Below are some possible ways of doing it with different pros & cons.

## A temporary solution

For logging and dataset configurations, or any future configurations sensitive to relative paths:

- Recursively traverse through the configuration dictionary
- Whenever a path value is encountered, denoted by certain keys, e.g. filepath, filename or path, expand the path to be absolute path by prepending project_path to it.

### Pros & Cons

* **Pros**: backward-compatible. Will work with not just logging config and datasets config, but also other configuration in the future that requires relative paths.
* **Cons**: it might produce bug that is very hard to trace in the future, although I'm not sure yet what that might be. It also makes the code less readable and adds some performance penalty to `load_context` due to the recursive loop. 

### The ideal solution

The ideal solution is to have a different execution model in which the knowledge of `project_path` is available globally for all components participating in the execution of kedro.