Unpin setuptools (#712)
Pinning setuptools's upper bound will cause conflict with dependencies in user's projects that don't pin it. For example, consider a user's project which has the following in their requirements.txt, which is very common:

```
ipython
kedro
```

Ipython also specifies setuptools as a dependency without an upper bound and it is always listed before Kedro in alphabetical order. When pip tries to resolve dependency, at the moment, since it only performs conflicts resolution in a "first-encounter wins" manner, it will register latest setuptools from IPython as a dependency, which will conflict with Kedro's pinned version. If user moves IPython below Kedro, it should work correctly again.

So to be absolutely certain, this commit unpins setuptools.