Add a video dataset (#1312)
This commit adds a video dataset that can read and write video files. The
dataset is backed by OpenCVs video reader and writer and different
buffer protocols such as PIL.Image and numpy.ndarray. There is one class
for iterables (e.g. list of frames) and one for generators. Since large
videos might not fit in memory, this ability allows us to read one frame
from the video file, transform the frame and write it to another video
file. Thus there is only need for one frame at a time in memory.

The different codecs that are supported depends on the codecs installed
and how OpenCV was compiled.

Signed-off-by: Daniel Falk <daniel.falk.1@fixedit.ai>
Signed-off-by: Nok <nok_lam_chan@mckinsey.com>