Add Prefect 2.0 guide for simple deployment with agent on localhost (#2748)
* Add Prefect 2.0 guide for simple deployment with agent

Add a step by step deployment guide for Kedro with Prefect 2.0. The guide includes a basic example of leveraging local resources with an agent that pulls from a work pool

Signed-off-by: Juan Lovera <8958924+jmalovera10@users.noreply.github.com>

* Include introduction section suggestions

Clarify on the conditions in which this guide was executed and better orient the developer to earlier versions for documentation on Prefect 1.0. Fix redaction on second paragraph.

Signed-off-by: Juan Lovera <8958924+jmalovera10@users.noreply.github.com>

---------

Signed-off-by: Juan Lovera <8958924+jmalovera10@users.noreply.github.com>
Co-authored-by: Jo Stichbury <jo_stichbury@mckinsey.com>